+++ functions/src/index.ts
@@ -25,7 +25,8 @@ initializeApp();
 // functions should each use functions.runWith({ maxInstances: 10 }) instead.
 // In the v1 API, each function can only serve one request per container, so
 // this will be the maximum concurrent request count.
-setGlobalOptions({ maxInstances: 10 });
+setGlobalOptions({ maxInstances: 10, region: 'southamerica-east1' });
 
 export { syncRoleClaim } from './syncRoleClaim';
 export { crearUsuario, editarUsuario, editarEmailUsuario, eliminarUsuario } from './gestionUsuarios';
+export { verificarVencimientosDocumentacion } from './verificarVencimientosDocumentacion';